Remove githubmanager/pr-watcher
