--- samples/config/pr-watcher-config.yaml
@@ -1,2 +1,0 @@
-spaces:
-  - name: public
