+++ test/unit/buildrunner/test_monitoring_interval_floor.py
@@ -19,10 +19,9 @@
 A 0 (or negative) interval turns the BuildWatcher poll loop and the dispatched
 BuildRunner event loop into CPU/storage busy-loops. The floor is enforced with a
 single declarative bound — ``Field(ge=MIN_MONITORING_INTERVAL_SECONDS)`` — on
-both watcher configs (BuildWatcherConfig drives the watcher's own poll loop and
-the runners it spawns; PrWatcherConfig drives the PR watcher's poll loop). The
-build-runner CLI applies the same bound via ``click.IntRange``. An out-of-range
-value is rejected at construction rather than silently clamped.
+BuildWatcherConfig, which drives the watcher's own poll loop and the runners it
+spawns. The build-runner CLI applies the same bound via ``click.IntRange``. An
+out-of-range value is rejected at construction rather than silently clamped.
 """
 
 import pytest
@@ -30,12 +29,11 @@ from pydantic import ValidationError
 
 from gbserver.types.buildwatcherconfig import BuildWatcherConfig
 from gbserver.types.constants import MIN_MONITORING_INTERVAL_SECONDS
-from gbserver.types.prwatcherconfig import PrWatcherConfig
 
 _FLOOR = MIN_MONITORING_INTERVAL_SECONDS
 
 
-@pytest.mark.parametrize("config_cls", [BuildWatcherConfig, PrWatcherConfig])
+@pytest.mark.parametrize("config_cls", [BuildWatcherConfig])
 @pytest.mark.parametrize("bad", [0, -1, -5])
 def test_monitoring_interval_rejected_at_construction(config_cls, bad):
     """A below-floor interval raises rather than silently busy-looping."""
@@ -43,7 +41,7 @@ def test_monitoring_interval_rejected_at_construction(config_cls, bad):
         config_cls(monitoring_interval=bad)
 
 
-@pytest.mark.parametrize("config_cls", [BuildWatcherConfig, PrWatcherConfig])
+@pytest.mark.parametrize("config_cls", [BuildWatcherConfig])
 def test_valid_monitoring_intervals_accepted(config_cls):
     assert config_cls().monitoring_interval == 5  # default unchanged
     assert config_cls(monitoring_interval=_FLOOR).monitoring_interval == _FLOOR
